refactor: move Android dependencies of the new previewer to another file
